Add config env.

+++ public/api/todos/todos.js
@@ -2,7 +2,6 @@ angular.module('myApp').controller('myCtrl', function ($scope, $http, $location,
     var self = this;
     self.mode = '';
     self.user = loggedin;
-    $('#id_alert').hide();
     $http({
         method: 'GET',
         url: 'http://localhost:3000/todos',
@@ -39,6 +38,8 @@ angular.module('myApp').controller('myCtrl', function ($scope, $http, $location,
         });
     }
     self.delete = function (id) {
+       if (confirm('do you want to delete this todo'))
+        {
         $http({
             method: 'DELETE',
             url: 'http://localhost:3000/todos/' + id,
@@ -57,12 +58,10 @@ angular.module('myApp').controller('myCtrl', function ($scope, $http, $location,
                     jj = jj + 1;
                 }
             }
-            $('#id_alert').hide();
-            self.idup = "";
-            self.titleup = "";
         }).error(function (data, status) {
             console.log("log in function delete in block error " + "status = " + status + "data =" + JSON.stringify(data));
         });
+        }
     }
     self.done = function (id, p) {
         var x = !(p.done);
@@ -124,18 +123,7 @@ angular.module('myApp').controller('myCtrl', function ($scope, $http, $location,
                 alert('error in server');
             }
         });
-
     };
-    self.ready_delete = function (id, title) {
-        self.idup = id;
-        self.titleup = title;
-        $('#id_alert').show();
-    }
-    self.hide_alert = function () {
-        $('#id_alert').hide();
-        self.idup = "";
-        self.titleup = " ;"
-    }
     self.setmode = function (mode, p) {
         self.clear();
         self.mode = mode;
